enhance(examples): improve examples and getting started
