+++ etl/steps/data/examples/examples/latest/private_example.py
@@ -1,9 +1,9 @@
 import pandas as pd
 from owid.catalog import Dataset, Table
+from owid.catalog.utils import underscore_table
 from owid.walden import Catalog
 
 from etl.steps.data.converters import convert_walden_metadata
-from owid.catalog.utils import underscore_table
 
 
 def run(dest_dir: str) -> None:
